Elaborate AST with proofs for type class constraints in inferencer
Currently, the type inferencer can prove that type class constraints are
satifiable but throws the proof away after that. Hence, the type checker and the
type class elimnator have to redo the proof.

This patch changes the type inferencer such that it leaves the constructed
proofs in AST (as values of type `Dict`). The type checker now checks that the
proof if valid and the type class eliminator proof `Dict` into an actual
dictionary value.