Make pukeko aware of ruceki
Split the compiler into several commands and add one to produce the
`NoLambda` code in JSON. Change a few examples to make them work in a strict
interpreter. Add a few tests for ruceki to the run-examples test suite. Add
a few benchmarks for ruceki as well.

Also, stop using stylish-haskell.